Added program to reverse an inputted linked list
Made a function to reverse a linked list. IM SO HAPPY THIS WORKED FIRST GO. Haven't had this kind of satisfaction in a long time. P.S. TA's if you're reading this, sorry for that massive comment block above the reverse function but I hope that makes it a bit clearer what im trying to achieve in the function.