Updated queue implementation program
Everything works now. Asks user for array size input, generates an array of that size with random values. Then generates a queue from the array values and prints it. Then generates a queue again but this time starting from the back of the array, basically a reversed version of the array is stored in the queue. All queue functions are done using convention of FIFO enqueueing and dequeueing. ON TO STACKS NOW HEHEHE