Created a file for binary search trees
Current program does: 1. Generates an array with random values of size 'n' where n is input by the user. 2. Generates a binary search tree from the array using the first value of the array as the root. TODO: Make the tree generate from the value in the array that is the closest to the 'mean' value.