Updated BST program See more in desc
The program can now 1. generate a balanced binary search tree from a randomly generated array of elements. Made new generate functions to make sure the BST is balanced. Older function still exists in here for a record of my thought process. 2. Search the tree for an element (See lines 240 and 257 to use this functionality)