Updated BST random functions program to delete elements from a tree
The program now generates a balanced binary search tree from an array of randomly generated elements. It then prints this tree and asks the user which element they want to delete. It then deletes (or doesnt depending if the element exists in the tree) and then prints the modified binary tree.