Created a full program with a menu to perform functions on a binary heap.
Program starts by asking the user to input the size of the heap and generates a heap. A Menu is then displayed with the following options: 1. Re-generate heap 2. Extract minimum value 3. Insert value 4. Display heap 5. Empty heap 6. Exit. All menu options are fully functional. I changed the way insert works from the last iteration and used the heapifyUp function inside it. Functions are all according to the specification in the docs file shared with us.