Created deletion from AVL Trees and made a menu to perform different AVL tree functions
Instead of balancing the trees in the insert/delete functions I made separate functions for insert balancing and deletion balancing. TODO: Currently 'emptying' the tree is just setting the root to NULL. I have to create a function to actually free each node and then set the root to NULL as the other nodes are still laid out in memory.